requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
schedule==1.2.1
numpy
pandas
python-dotenv==1.0.1
requests==2.31.0
ollama==0.1.6